Unicast accept refactoring
