+++ zenoh/src/net/transport/unicast/mod.rs
@@ -81,10 +81,10 @@ pub(crate) struct TransportConfigUnicast {
     pub(crate) peer: PeerId,
     pub(crate) whatami: WhatAmI,
     pub(crate) sn_resolution: ZInt,
-    pub(crate) initial_sn_tx: ZInt,
-    pub(crate) initial_sn_rx: ZInt,
     pub(crate) is_shm: bool,
     pub(crate) is_qos: bool,
+    pub(crate) initial_sn_tx: ZInt,
+    pub(crate) initial_sn_rx: ZInt,
 }
 
 /// [`TransportUnicast`] is the transport handler returned
@@ -94,7 +94,7 @@ pub struct TransportUnicast(Weak<TransportUnicastInner>);
 
 impl TransportUnicast {
     #[inline(always)]
-    pub(super) fn get_transport(&self) -> ZResult<Arc<TransportUnicastInner>> {
+    pub(super) fn get_inner(&self) -> ZResult<Arc<TransportUnicastInner>> {
         self.0
             .upgrade()
             .ok_or_else(|| zerror!("Transport unicast closed").into())
@@ -102,42 +102,42 @@ impl TransportUnicast {
 
     #[inline(always)]
     pub fn get_pid(&self) -> ZResult<PeerId> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.get_pid())
     }
 
     #[inline(always)]
     pub fn get_whatami(&self) -> ZResult<WhatAmI> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.get_whatami())
     }
 
     #[inline(always)]
     pub fn get_sn_resolution(&self) -> ZResult<ZInt> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.get_sn_resolution())
     }
 
     #[inline(always)]
     pub fn is_shm(&self) -> ZResult<bool> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.is_shm())
     }
 
     #[inline(always)]
     pub fn is_qos(&self) -> ZResult<bool> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.is_qos())
     }
 
     #[inline(always)]
     pub fn get_callback(&self) -> ZResult<Option<Arc<dyn TransportPeerEventHandler>>> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport.get_callback())
     }
 
     pub fn get_peer(&self) -> ZResult<TransportPeer> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         let tp = TransportPeer {
             pid: transport.get_pid(),
             whatami: transport.get_whatami(),
@@ -154,7 +154,7 @@ impl TransportUnicast {
 
     #[inline(always)]
     pub fn get_links(&self) -> ZResult<Vec<Link>> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         Ok(transport
             .get_links()
             .into_iter()
@@ -164,14 +164,14 @@ impl TransportUnicast {
 
     #[inline(always)]
     pub fn schedule(&self, message: ZenohMessage) -> ZResult<()> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         transport.schedule(message);
         Ok(())
     }
 
     #[inline(always)]
     pub async fn close_link(&self, link: &Link) -> ZResult<()> {
-        let transport = self.get_transport()?;
+        let transport = self.get_inner()?;
         let link = transport
             .get_links()
             .into_iter()
@@ -186,7 +186,7 @@ impl TransportUnicast {
     #[inline(always)]
     pub async fn close(&self) -> ZResult<()> {
         // Return Ok if the transport has already been closed
-        match self.get_transport() {
+        match self.get_inner() {
             Ok(transport) => transport.close(tmsg::close_reason::GENERIC).await,
             Err(_) => Ok(()),
         }
@@ -199,7 +199,7 @@ impl TransportUnicast {
 
     #[cfg(feature = "stats")]
     pub fn get_stats(&self) -> ZResult<TransportUnicastStats> {
-        Ok(self.get_transport()?.stats.snapshot())
+        Ok(self.get_inner()?.stats.snapshot())
     }
 }
 
@@ -219,7 +219,7 @@ impl PartialEq for TransportUnicast {
 
 impl fmt::Debug for TransportUnicast {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self.get_transport() {
+        match self.get_inner() {
             Ok(transport) => f
                 .debug_struct("Transport Unicast")
                 .field("pid", &transport.get_pid())
