+++ zenoh/src/net/transport/unicast/transport.rs
@@ -51,14 +51,8 @@ macro_rules! zlinkindex {
 /*************************************/
 #[derive(Clone)]
 pub(crate) struct TransportUnicastInner {
-    // The manager this channel is associated to
-    pub(super) manager: TransportManager,
-    // The remote peer id
-    pub(super) pid: PeerId,
-    // The remote whatami
-    pub(super) whatami: WhatAmI,
-    // The SN resolution
-    pub(super) sn_resolution: ZInt,
+    // Transport config
+    pub(super) config: TransportUnicastConfig,
     // Tx conduits
     pub(super) conduit_tx: Arc<[TransportConduitTx]>,
     // Rx conduits
@@ -69,13 +63,12 @@ pub(crate) struct TransportUnicastInner {
     pub(super) callback: Arc<RwLock<Option<Arc<dyn TransportPeerEventHandler>>>>,
     // Mutex for notification
     pub(super) alive: AsyncArc<AsyncMutex<bool>>,
-    // The transport can do shm
-    pub(super) is_shm: bool,
     // Transport statistics
     #[cfg(feature = "stats")]
     pub(super) stats: Arc<TransportUnicastStatsAtomic>,
 }
 
+#[derive(Clone)]
 pub(crate) struct TransportUnicastConfig {
     pub(crate) manager: TransportManager,
     pub(crate) pid: PeerId,
@@ -135,16 +128,12 @@ impl TransportUnicastInner {
         }
 
         TransportUnicastInner {
-            manager: config.manager,
-            pid: config.pid,
-            whatami: config.whatami,
-            sn_resolution: config.sn_resolution,
+            config,
             conduit_tx: conduit_tx.into_boxed_slice().into(),
             conduit_rx: conduit_rx.into_boxed_slice().into(),
             links: Arc::new(RwLock::new(vec![].into_boxed_slice())),
             callback: Arc::new(RwLock::new(None)),
             alive: AsyncArc::new(AsyncMutex::new(true)),
-            is_shm: config.is_shm,
             #[cfg(feature = "stats")]
             stats: Arc::new(TransportUnicastStatsAtomic::default()),
         }
@@ -163,7 +152,7 @@ impl TransportUnicastInner {
     /*           TERMINATION             */
     /*************************************/
     pub(super) async fn delete(&self) -> ZResult<()> {
-        log::debug!("Closing transport with peer: {}", self.pid);
+        log::debug!("Closing transport with peer: {}", self.config.pid);
 
         // Mark the transport as no longer alive and keep the lock
         // to avoid concurrent new_transport and closing/closed notifications
@@ -177,7 +166,11 @@ impl TransportUnicastInner {
         }
 
         // Delete the transport on the manager
-        let _ = self.manager.del_transport_unicast(&self.pid).await;
+        let _ = self
+            .config
+            .manager
+            .del_transport_unicast(&self.config.pid)
+            .await;
 
         // Close all the links
         let mut links = {
@@ -209,7 +202,7 @@ impl TransportUnicastInner {
         let guard = zread!(self.links);
         #[cfg(feature = "transport_multilink")]
         {
-            if guard.len() < self.manager.config.unicast.max_links {
+            if guard.len() < self.config.manager.config.unicast.max_links {
                 zlinkget!(guard, link).is_none()
             } else {
                 false
@@ -226,7 +219,7 @@ impl TransportUnicastInner {
             bail!(
                 "Can not add Link {} with peer {}: max num of links reached",
                 link,
-                self.pid,
+                self.config.pid,
             )
         }
 
@@ -257,7 +250,11 @@ impl TransportUnicastInner {
                 Ok(())
             }
             None => {
-                bail!("Can not start Link TX {} with peer: {}", link, self.pid)
+                bail!(
+                    "Can not start Link TX {} with peer: {}",
+                    link,
+                    self.config.pid
+                )
             }
         }
     }
@@ -270,7 +267,11 @@ impl TransportUnicastInner {
                 Ok(())
             }
             None => {
-                bail!("Can not stop Link TX {} with peer: {}", link, self.pid)
+                bail!(
+                    "Can not stop Link TX {} with peer: {}",
+                    link,
+                    self.config.pid
+                )
             }
         }
     }
@@ -283,7 +284,11 @@ impl TransportUnicastInner {
                 Ok(())
             }
             None => {
-                bail!("Can not start Link RX {} with peer: {}", link, self.pid)
+                bail!(
+                    "Can not start Link RX {} with peer: {}",
+                    link,
+                    self.config.pid
+                )
             }
         }
     }
@@ -296,7 +301,11 @@ impl TransportUnicastInner {
                 Ok(())
             }
             None => {
-                bail!("Can not stop Link RX {} with peer: {}", link, self.pid)
+                bail!(
+                    "Can not stop Link RX {} with peer: {}",
+                    link,
+                    self.config.pid
+                )
             }
         }
     }
@@ -330,7 +339,11 @@ impl TransportUnicastInner {
                     Target::Link(stl.into())
                 }
             } else {
-                bail!("Can not delete Link {} with peer: {}", link, self.pid)
+                bail!(
+                    "Can not delete Link {} with peer: {}",
+                    link,
+                    self.config.pid
+                )
             }
         };
 
@@ -346,23 +359,23 @@ impl TransportUnicastInner {
     /*            ACCESSORS              */
     /*************************************/
     pub(crate) fn get_pid(&self) -> PeerId {
-        self.pid
+        self.config.pid
     }
 
     pub(crate) fn get_whatami(&self) -> WhatAmI {
-        self.whatami
+        self.config.whatami
     }
 
     pub(crate) fn get_sn_resolution(&self) -> ZInt {
-        self.sn_resolution
+        self.config.sn_resolution
     }
 
     pub(crate) fn is_shm(&self) -> bool {
-        self.is_shm
+        self.config.is_shm
     }
 
     pub(crate) fn is_qos(&self) -> bool {
-        self.conduit_tx[0].priority != Priority::default()
+        self.config.is_qos
     }
 
     pub(crate) fn get_callback(&self) -> Option<Arc<dyn TransportPeerEventHandler>> {
@@ -373,7 +386,7 @@ impl TransportUnicastInner {
     /*           TERMINATION             */
     /*************************************/
     pub(crate) async fn close_link(&self, link: &LinkUnicast, reason: u8) -> ZResult<()> {
-        log::trace!("Closing link {} with peer: {}", link, self.pid);
+        log::trace!("Closing link {} with peer: {}", link, self.config.pid);
 
         let guard = zread!(self.links);
         if let Some(l) = zlinkget!(guard, link) {
@@ -384,7 +397,7 @@ impl TransportUnicastInner {
             // Schedule the close message for transmission
             if let Some(pipeline) = pipeline.take() {
                 // Close message to be sent on the target link
-                let peer_id = Some(self.manager.pid());
+                let peer_id = Some(self.config.manager.pid());
                 let reason_id = reason;
                 let link_only = true; // This is should always be true when closing a link
                 let attachment = None; // No attachment here
@@ -401,7 +414,7 @@ impl TransportUnicastInner {
     }
 
     pub(crate) async fn close(&self, reason: u8) -> ZResult<()> {
-        log::trace!("Closing transport with peer: {}", self.pid);
+        log::trace!("Closing transport with peer: {}", self.config.pid);
 
         let mut pipelines: Vec<Arc<TransmissionPipeline>> = zread!(self.links)
             .iter()
@@ -409,7 +422,7 @@ impl TransportUnicastInner {
             .collect();
         for p in pipelines.drain(..) {
             // Close message to be sent on all the links
-            let peer_id = Some(self.manager.pid());
+            let peer_id = Some(self.config.manager.pid());
             let reason_id = reason;
             // link_only should always be false for user-triggered close. However, in case of
             // multiple links, it is safer to close all the links first. When no links are left,
@@ -431,10 +444,10 @@ impl TransportUnicastInner {
     pub(crate) fn schedule(&self, mut message: ZenohMessage) {
         #[cfg(feature = "shared-memory")]
         {
-            let res = if self.is_shm {
+            let res = if self.config.is_shm {
                 message.map_to_shminfo()
             } else {
-                message.map_to_shmbuf(self.manager.shmr.clone())
+                message.map_to_shmbuf(self.config.manager.shmr.clone())
             };
             if let Err(e) = res {
                 log::trace!("Failed SHM conversion: {}", e);
